feature/add-domain-services: fix common layers for posts.

+++ src/Common/InstaConnect.Common.Application/Abstractions/ICollectionQueryResponse.cs
@@ -6,7 +6,7 @@ public interface ICollectionQueryResponse
 
     public int PageSize { get; }
 
-    public int TotalCount { get; }
+    public long TotalCount { get; }
 
     public bool HasNextPage { get; }
